+++ src/main/java/at/porscheinformatik/seleniumcomponents/clarity/ClarityInputContainerComponent.java
@@ -62,7 +62,7 @@ public class ClarityInputContainerComponent extends ClarityFormControlContainer
     public static ClarityInputContainerComponent byLabel(SeleniumComponent parent, String label)
     {
         String xpath = String.format(
-            ".//clr-input-container[.//label[contains(@class, 'clr-control-label') and contains(text(),'%s')]]", label);
+            ".//clr-input-container[.//label[contains(@class, 'clr-control-label') and contains(.,'%s')]]", label);
 
         return new ClarityInputContainerComponent(parent, WebElementSelector.selectByXPath(xpath));
     }